vcc: New $Alias stanza for VMODs
It is now possible to rename a symbol from a VMOD without breaking
existing code until the VMOD author decides to remove the deprecated
alias.

The VCC file syntax is simply:

    $Alias <alias> <function>

    [Optional description]

    $Alias <.alias> <obj.method>

    [Optional description]

The alias can only apply to a $Function or $Method.

The generated RST looks like this for a method:

    ALIAS x<object>.<name>()
    ------------------------

    Deprecated alias for ``x<object>.<method>()``.

    [Optional description]

It looks the same for a function without the $Object prefix. Aliases
don't have a reference label for sphinx docs, the goal is not to make
them prominent, but rather to have the ability to move them to a manual
section for deprecated symbols at the end of the document.