Respect TOSTRAND() lifetime
In a different context, I questioned my understanding of C object
lifetimes and learned that I had given the wrong explanation for why
TOSTRAND() was used wrongly in
0c96fc6597c3ba6da057bb43acb7f677e6b776f8:

TOSTRAND() creates a compound literal whose lifetime is the enclosing
block.

Thus, the error was unrelated to temporary lifetime, but rather that
the compound literal was assigned in the if-block, but used outside
it.

Ref: https://stackoverflow.com/questions/71225472/more-questions-on-the-c11-temporary-lifetime-rule-and-undefined-behaviour/71226049#71226049