req.body: New req_body_partial flag for caching
Partial caching allow VMODs to fetch and inspect part of the body. This
gives the same properties as req_body_cached but allows the request body
to be larger than what is stored. If a bereq send failure happens within
the partial cached body delivery, it also becomes eligible to retries.

This enables a more robust caching and retry policy where not knowing
the complete body size (chunked or h2 without content-length) or on the
other hand knowing that a body is larger (content-length) does not lead
to VCL failures.

It can also be valuable to inspect the beginning of a request body. For
example an image upload service could parse images metadata directly in
Varnish to enforce rules such as specific file formats or how large in
pixels an image may be.

Better diff with the --ignore-all-space option.