V1F: Add state and readahead buffer for v1f chunked processing
This commit contains only the init/fini mechanics for the
state/buffer, it is not used yet.

The state is allocated on the heap to be prepared for partial caching
(see #3798), where the V1F layer is left and control returns to VCL:

- the worker workspace aws is not an option because it must be freed
  to its original state before returning to VCL.

- the task (req) workspace is not an option because of rollbacks.

While I would have preferred to avoid a heap allocation, it will be
payed off by the reduction of system calls in follow up commits.