V1F: Refactor v1f_read() to use scatter/gather reads (readv())
This is to support optional readaheads:

The first io vector is for read data which the caller absolutely
needs, any other io vectors are for potential readahead.

The readahead is not used yet, for now, all v1f_read() calls
use the IOV() macro to just wrap the existing single buffer.