transit buffer without the timeout on the condition variable
We do not need to periodically check the conditions in
obj_extend_condwait() if we ensure that HSH_Cancel() triggers a
wakeup.

Readers on private objects need to ensure they call HSH_Cancel() if
they abort the read, so this is consistent with our requirements.

While we could add a seperate function to the object API for the sole
purpose of signaling a cancel, HSH_Cancel() already calls
ObjWaitState() to syncronize with the backend thread, so adding the
signal there was deemed the simpler solution.

As agreed during bugwash, Martin wants to add back the timeout
as an optinal parameter (default: no timeout).