param: Rename idle_send_timeout to resp_idle_interrupt
Exceptionally, one existing VTC was migrated to the new name immediately
since it relied on debug logs containing 'idle_send_timeout'.

Turning the timeout to an interrupt, I tried to update the parameter
description to better convey what it does. This does not change the
current behavior, but I tried to highlight what is wrong with this
timeout and the mismatch between its purpose for HTTP/1 and h2.

Depending on where we look at the documentation, between the parameter
or VCL variables, we get a slightly different description. None of them
are accurate, but the consensus seems to have always been the parameter
description from before h2 or the 'sess.idle_send_timeout' variable.

For session-level timeouts, we may consider an h2_send_timeout parameter
in the future to distinguish between socket writes and underlying h2
streams timing out on a condvar. This h2_send_timeout parameter would
map to SO_SNDTIMEO for h2 client connections.

Refs #2927