smd: Plug inconsequential leak
The debug stevedore only acts as a wrapper around the malloc storage, so
it proxies its initialization method. The allocation of av isn't tracked
so it leaks when it runs out of scope, along with the copies of values
assigned to the array.

Since a reference to aav is kept in stevedore::av, we can safely pass
direct references to arguments to the underlying malloc storage instead
of copies.

Spotted by asan.