varnishncsa: Add support for the -k option
I found myself needing this option to incrementally build a query and a
format on a large VSL compressed file. Because of the small nature of
the ncsa output I was working with and the narrow query I was refining,
and the decompression delay, it would take forever to complete this
pipeline:

    zcat '...' | varnishncsa -r - -g request -q '...' -F '...' | head -n 5

With the -k option it happens instantly. Alternatively, it could
probably have worked also with the -u option, but it wouldn't be a
one-liner change, and it would also result in a more complex command:

    zcat '...' | varnishncsa -r - -g request -q '...' -F '...' -k 5
    # vs
    zcat '...' | varnishncsa -r - -g request -q '...' -F '...' -u | head -n 5

The -k option would also generally be useful in "live" mode.