varnishstatdiff: New utility to compare metrics
As I was comparing the output of two varnishstat executions that were
captured after a fresh start of varnishd followed by the workloads to
compare, I realized diff(1) was giving me a hard time, and git-diff(1)
barely improved the situation.

Looking for generic command line utilities to compare metrics I wasn't
able to find anything. So instead I came up with the output format I
thought would help me spot interesting differences and came up with a
format inspired by the unified diff, with a twist. I wanted metrics to
be vertically aligned to easily see differences in orders of magnitude
and reduce the noise to a minimum.

The result is that taking this detour to script varnishstatdiff sped
my research up ultimately.

This should hopefully be portable to POSIX systems.