task: Add a TASK_QUEUE_RESERVE macro
The new macro is added to better reflect operations that cater to the
thread reserve. It also enables the allocation the right number of
reserve heads in struct pool, instead of having one for each priority.
It grants the possibility of priority classes so low that they wouldn't
be eligible to queuing, without wasting space with needless queue heads.