wrk: Give bgfetch tasks the lowest priority
The goal is to prevent grace mode from adding load to a saturated
Varnish server.

A background fetch entering the queue will block the client task that
triggerred it until it starts its execution, and reaches the point
where it no longer needs to hold onto its req.

On a saturated system this can result in significant client latency
despite a grace hit. A stale object can be served until the end of its
grace period, at which point a regular fetch would be attempted, and
eligible to queuing if that is still necessary.

This turns the Pool_Task() failure dead branch from VBF_Fetch() into a
reachable one.