cocci: Make isomorphisms actually work
It turns out declaring typedefs in coccinelle works better with the
typedef keyword. Reusing the same keyword for the same purposes seems
so obvious in hindsight.

What happened before was that a generic type would be created, making
VCL_BOOL or VCL_VOID match virtually any type, completely defeating
the isomorphism purpose. With this out of the way, we can actually
resume the effort of adding more typedefs.

To summarize:

- we should declare a typedef as an isomorphism
- we should define macros as such

I'm considering renaming tools/coccinelle/vdef.h to varnish.h to match
the varnish.iso file name and wrapping the README instructions in a new
shell script.