Add a workspace usage reporting facility
We add tracking of a "high water" pointer to workspaces.

It is tracked automatically for normal allocations and is preseved
across resets.

For reservations, the used space needs to be tracked explicitly.
If explicit tracking is missing, the "at least" flag is raised
if the reserved space exceeded the high water mark.

This could be done more intelligently, "at least" could be cleared
if a following allocation exceeds a previous reservation, but
the way to go is to have proper reporting.

For user reporting, a prototypical VSL is implemented as SLT_Debug, for
example:

**** v1    vsl|       1001 Debug           c ws highwater client = 77684
**** v1    vsl|       1001 Debug           c ws highwater session = 120
**** v1    vsl|       1001 Debug           c ws highwater thread = 2360

With the "at least" flag present, = is replaced by >= as in this example:

*** v1    vsl|       1027 Debug           c ws highwater client >= 77761

To avoid struct ws to grow by another pointer size, the magic value
is reduced to a uint16_t to make room for the flags.