varnishtest: txresp Date header by default
This commit prepares varnishtest for a change to Varnish revalidations,
where the Last-Modified header must be at least one second older than
the Date header for Varnish to send an If-Modified-Since header.

Any explicitly defined Date header will override the default, and
the Date header can be omitted with -nodate.