Drop ban_mtx during STV_BanExport()
Stevedores now need to be prepared to potentially receive multiple
export calls in parallel.

Why is this change safe?

ban_export() needs to hold the lock while serializing the ban list
into the VSB, but once that is done, it can drop the lock.

Other call sites are:

* BAN_Commit() -> ban_info_new()

  Because our new ban b could be concurrently removed by another
  thread while the lock is dropped, we move the dup check to before
  the ban_info_new() call.

  This should also have the added benefit of potentially (not)
  exporting more bans as completed.

* ban_cleantail() -> ban_info_drop()

  Each iteration of the loop starts with the last ban, and the now
  removed ban b has been moved to the freelist, so there is no
  concurrent access possible any more while the lock is dropped.

* BAN_Compile(), BAN_Shutdown()

  Called during init / fini, no concurrency here.

Other than that, we should really batch ban_export calls, especially
for ban_cleantail...