vslc: Plug mmap file descriptor leak
To be consistent with how the file cursor behaves, the close_fd field is
duplicated in the mmap cursor. If a VUT replaces stdin's file descriptor
with a regular file's fd using dup2(2), we don't want to close it just
because we managed to mmap(2) it.

For some reason we don't use the closefd() macro in the VSL cursor code,
potentially to avoid its underlying assertion in libvarnishapi.

On the other hand we do use it in other places:

    $ git grep -l closefd -- lib/libvarnishapi/
    lib/libvarnishapi/daemon.c
    lib/libvarnishapi/vsm.c

So maybe in a subsequent change `(void)close(fd)` statements could turn
into `closefd(&fd)` in vsl_cursor.c to harden those code paths as well.