Implement HTTP request and response egress validation
This implements HTTP character set validation run on the `struct http`
after vcl_deliver, vcl_synth and vcl_backend_fetch. The idea is to pick up
if the VCL program has ended up filling in illegal values that would
violate the HTTP protocol.

If a problem is found, we will error out with a 503. If the error is
caught in vcl_synth, the connection will be closed.