shmlog: Avoid allocating vsl_reclen bytes in VSLv()
When vsl_reclen increases to a large value, like 8kB to hold long ReqURL
or *Header:Cookie records, making such an allocation on the stack could
jump over a 4kB guard page. The underlying printf call could also result
in significant stack usage. Both factors could result in hard to debug
segmentation faults.

Instead, we can opportunistically print to a reasonably-sized buffer,
and fall back to printing directly inside the VSL segment when a record
is too large for the buffer. At this point we know the full record size
and we can adjust it to vsl_reclen.

This is inspired by a different patch from @mbgrydeland targeting the
6.0 branch.