vgz: Declare bit fields in the correct order
And fix the comment of the last_bit field, that for some reason does not
really match its meaning as if the "last" bit (pun intended) of its name
was confusingly interpreted as "the last bit position of the compressed
payload", which actually describes the stop_bit field instead.

The description of the stop_bit field is accurate, but may not reflect
that there may be a gap between the last bit of the block payload and
the following CRC32 byte. There may be dead bits if the compressed block
size is not a factor of 8.