gzip: We can't work with a standalone zlib
Unlike the inflate() API that lands itself nicely to figuring block
boundaries, I have come to the conclusion that this is impossible for
its deflate() counterpart.

I proceeded similarly to inflate():

- find how and when boundaries are registered
- port the formula to varnishd's code
- see what boundaries we operate on outside zlib

What I found is that we keep track of the start, stop and last bit
positions deep in the inflate code. The start bit is not a problem for
us, it is guaranteed to always be 80 bits, the initial 10B gzip header.
We know for a fact that gzip streams we generate never have anything
extra. This is verified by an assertion.

Porting the formula was trivial using the deflatePending() function
and it turns out to be essential to figure boundaries out.

The problem I initially suspected was that one call to deflate() could
result into multiple blocks being output, as the structure of the code
suggests in the trees.c file. This poor man's tracing confirms that we
can't achieve from the outside what our patched zlib does.

Output example:

    $ ./varnishtest -i tests/c00034.vtc | grep z= | sed 's/.*said//'
    z_deflate z=0x7fca7d016040 start=80
    VGZ_UpdateObj z=0x7fca7d016040 start=80 stop=0 last=0 bits=80
    z__tr_flush_block z=0x7fca7d016040 last=80
    z__tr_flush_block z=0x7fca7d016040 stop=196
    VGZ_UpdateObj z=0x7fca7d016040 start=80 stop=196 last=80 bits=264
    VGZ_UpdateObj z=0x7fca7d016040 start=80 stop=196 last=80 bits=264

We know that we can hard-code the start bit to 80, and we could maybe
capture the last bit (80) but unlike the inflate() code there doesn't
seem to be state we could inspect to inform this need. On the inflate()
side we can rely on the z_stream::data_type field, but here it seems to
solely contain Z_BINARY or Z_TEXT (if not Z_UNKNOWN) and no special bits
like in inflate mode.

After adding the Z_BLOCK flag in this patch, the situation does not
improve:

    $ ./varnishtest -i tests/c00034.vtc | grep z= | sed 's/.*said//'
    z_deflate z=0x7ff0e6416040 start=80
    VGZ_UpdateObj z=0x7ff0e6416040 start=80 stop=0 last=0 bits=196
    z__tr_flush_block z=0x7ff0e6416040 last=196
    z__tr_flush_block z=0x7ff0e6416040 stop=220
    VGZ_UpdateObj z=0x7ff0e6416040 start=80 stop=220 last=196 bits=288
    VGZ_UpdateObj z=0x7ff0e6416040 start=80 stop=220 last=196 bits=288

First, we get a chance to run VGZ_UpdateObj() after the initial header
and before the first block is written. So we can first record the start
at 80, so we gain nothing. Once again, we get a chance to capture the
last bit at 196 (instead of 80) but have no state to inspect to confirm
whether we should do that. By the time we come back the stop bit has
already been recorded and we already went past it. We also can't infer
the stop bit of the stream as `total_out * 8 - 64` (where 64 is the
number of bits needed for the CRC32 and ISIZE fields at the end of a
block) because the compressed payload of a block may not have a bit size
that is a multiple of 8.

Even worse, we artifically force the block boundary and end up with a
larger gzip stream altogether.

And that is just a simplistic test case when it comes to gzip. Another
one like v00004.vtc shows an even bleaker picture. And if we aren't even
close to making a gzip test case pass, there's no hope for ESI assembly
of multiple fragments compressed (or not) independently.