libvgz: Add a new z_bounds type
Its purpose is to collect block boundaries without changing the z_stream
layout, at the expense of a larger internal state. This is done next to
the tracking of block boundaries already present via z_stream, but this
is only transitional. It comes with a small API to collect boundaries
from outside of zlib.