build: Require an explicit --with-jemalloc to use it
This removes the automatic check made only on Linux systems for jemalloc
now that operating systems we care about ship a much more recent version
than the 3.6 that is known to work really well with Varnish on x86_64
hardware.

Of the platforms we provide packages for, only Ubuntu 18.04 (bionic) and
RHEL 7 (via EPEL) ship jemalloc 3.6.0 and the rest already moved on to a
5.x series.

It appears that jemalloc 3.6 on aarch64 results in unstable Varnish
workloads while jemalloc 5 is known to generate a lot of waste with its
default configuration with a highly threaded workload like Varnish would
operate, based on feedback we have seen over years.

From now on, jemalloc is found via pkg-config, only if it is explicitly
requested at configure time. With explicit opt-in, the Linux-only check
is gone and we solely rely on pkg-config to find the library. This can
of course be overriden as usual at configure time:

    ./configure --with-jemalloc JEMALLOC_CFLAGS=... JEMALLOC_LIBS=...

Or, if you don't like long command lines, via environment variables.

Refs #3867
Refs #3879