heritage: Remove unset params::[gu]id fields
They are shared with the cache process but are never used. Only the VCC
process uses them, but they are never set. This specific fchown(2) call
in the VCC process was probably a no-op in the first place: since the
fields are never set this is transferring ownership to root:root and if
that succeeded the process was already root in the first place. If it
failed, we never see the error message since we lacked root privileges.

Both the unix and solaris jails are designed to run VCC (and CC) with
limited privileges, and in the absence of a jail, the outcome should be
the same: VCC creates a file with credentials suitable for the next CC
invocation.