trf: Add client traffic management to the CLI
The interface offered for operations is a set of traffic.accept,
traffic.refuse and traffic.status commands with similar semantics
as start, stop and status.

The child process will eventually listen to the traffic.start and
traffic.refuse commands, but for now only a pseudo accept_traffic
parameter is added in struct params. It will be used when tasks
running on worker threads reach certain points to decide whether
to keep processing requests or not.