backend: queue tasks when no connections available
This patch allows a task to be queued when a backend reaches its
max_connections.  The task will queue on the backend and wait for a
connection to become availble, rather than immediately failing.

This initial commit just adds the basic functionality.  It temporarily
uses the connect_timeout as the queue wait time, until new parameters
are added in followup effort.