Improve the Unix jail documentation
Document that the daemon and worker user have to share their primary
group.

Provide an exmaple of how to set up a system for the default users.

Note: I am well aware of the commands in pkg-varnish-cache, but they
use Linux specific useradd syntax. The commands given hopefully are
portable - I tested them on Solaris and Linux.