build: Turn dot into a "proper" maintainer program
The --enable-maintainer-mode option has been used in autogen.des for a
while now and it has allowed automatic reconfiguration and hot reload of
makefiles.

For the next step, we can turn certain programs into maintainer programs
and only require them when they are absolutely needed. This is what we
currently do with SVG files generated from dot files, checked in git and
optionally required, with an error message when rebuild is not possible.

It turns out we only need 3 lines in configure.ac to implement all that,
with the following differences:

- no more `./configure --with-dot[=...]` option
  - use `./configure DOT=...` to override the default `dot` command
  - or alternatively run configure with `DOT=...` in the environment
- generic missing error message (but referencing dot) from make
- no need to reconfigure once `dot` is in the PATH

The DOT override is actually documented in `./configure --help` like the
`--with-dot[=...]` option was.

While dot is our most emblematic maintainer tool, with its build targets
checked in all the way in the git repository, this could be applied to
other maintainer programs like sphinx-build and rst2man that should not
be required when building from a dist archive.