build: Install vmod_*.vcc files
This change adds a new /usr/share/varnish/vcc (or equivalent) directory
where the built-in VMOD descriptors are installed. Only vmod_debug.vcc
is omitted since vmod_debug is not installed. Shipping VCC files can be
useful to end users, for example, to implement VCL autocompletion in a
text editor.

This change does not even begin to manage this for third-party VMODs.