Add individual hyperlink targets for vcl variables
in order to improve references to the documentation.

With this change, VCL variable documentation can be referenced
externally by replacing the dot in the variable name with a hyphen and
use that as a target, for example:

https://varnish-cache.org/docs/trunk/reference/vcl-var.html#resp-filters

Rules:

- only current (VCL >= 4.1) variables gain a target

- if there is a section by the same name as a target, do not change it

- No .* for HEADERS like req.http.*

Notes:

This turned out to be more work than anticipated for presumably a
minor change. First I tried to avoid adding target names by turning
the definition list items into subsections, which, with a simple
change to generate.py, worked ok for HTML, but looked terrible for the
man page, which only has three section levels: The variables appeared
on the same indentation level as the section headers.

So next I tried writing code to perform the change automatically, but
this also turned out to be tedious. So ultimately, I used an emacs
macro with manual polishing and the work was done in minutes.

Code is not always the answer...