stevedore: Log failures to create objects
We only log a Storage record when we successfully create an object, but
there may be no clue regarding which storage backend failed to allocate.
We can infer from stevedore VSCs where allocation failures happened, but
knowing from a VCL transaction which one failed will give a definitive
answer.

This is logged as an Error record, and the existing FetchError record
from VFPs ("Could not get storage") is left alone.