smug: Pass file descriptors from mgt to cache
The new smuggling facility allows mgt to abuse the heritage process and
sneak file descriptors to the cache process via a socket pair. This new
trade route will allow the cache to escape its jail and get privileged
file descriptors anyway. The smuggler's fence is hidden in plain sight,
in the heritage facility, which explains such a smug attitude.

Signed-off-by: Dridi Boukelmoune <dridi.boukelmoune@gmail.com>