VCC: ignore ENOENT on cleanup if vcc fails
When VCC is run, some temporary files are generated during the
process and are cleaned up at the end. The cleanup tries to remove
all the files that are supposed to be generated and displays a
"Could not delete 'vcl_...': No such file or directory" message if it
fails to find one of them. The error happens when VCC fails before
generating the file(s). This commit avoids the bogus messages by
ignoring ENOENT if VCC failed.

Fixes #3925